+++ src/app/pages/subjects/web/url-card/url-card.component.ts
@@ -6,6 +6,7 @@ import {NbToastrService, NbDialogService} from "@nebular/theme";
 import {WindowEditComponent} from "../../../articles/window-edit/window-edit.component";
 import {GeneralService} from "../../../../@core/utils/general.service";
 import {environment} from "../../../../../environments/environment";
+import {DeleteConfirmationComponent} from "../../delete-confirmation/delete-confirmation.component";
 
 @Component({
   selector: 'ngx-url-card',
@@ -19,13 +20,14 @@ export class UrlCardComponent implements OnInit {
   @Input() post: Post;
   icon: string;
   showOverlay: boolean = false;
+  deletingArticle: boolean = false;
 
   constructor(
     private photosService: PhotosService,
     private postService: PostService,
     private toastr: NbToastrService,
     private dialogService: NbDialogService,
-    private generalService: GeneralService
+    private generalService: GeneralService,
 
   ) {
     this.icon = this.subject === "MME" ? 'play-circle-outline' : 'external-link-outline';
@@ -35,17 +37,26 @@ export class UrlCardComponent implements OnInit {
   }
 
   deleteArticle(id) {
-    this.generalService.delete(environment.models.article, id).subscribe(result => {
-      if (result) {
-        this.loadVideos.emit();
-        this.toastr.success('Příspěvek byl úspěšně smazán!', 'Smazáno');
+
+    this.dialogService.open(DeleteConfirmationComponent).onClose.subscribe(state => {
+      if(state) {
+        this.deletingArticle = true;
+        this.generalService.delete(environment.models.article, id).subscribe(result => {
+          if (result) {
+            this.loadVideos.emit();
+            this.toastr.success('Příspěvek byl úspěšně smazán!', 'Smazáno');
+            this.deletingArticle = false;
+          }
+        })
+      } else {
+        this.deletingArticle = false;
       }
     })
+
   }
 
   editArticle(article) {
     this.dialogService.open(WindowEditComponent, {context: {article: article}}).onClose.subscribe(() => {
-
       this.loadVideos.emit()
     });
   }
